+++ app/hooks/useBalance.ts
@@ -1,4 +1,4 @@
-import { useContext } from "react";
+import { useContext, useEffect, useState } from "react";
 
 import Balance from "../models/schemas/BalanceSchema";
 import Category from "../models/schemas/CategorySchema";
@@ -6,14 +6,21 @@ import BalanceHistoryItemInterface from "../models/interfaces/BalanceHistoryItem
 
 import { RealmContext } from "../configs/RealmContext";
 import { AppConfigContext } from "../context/appConfig.context";
+import { Results } from "realm/dist/bundle";
 
 const { useQuery } = RealmContext;
 
 const useBalance = () => {
   const listCategory = useQuery(Category);
-  const listBalance = useQuery(Balance);
+  const fullListBalance = useQuery(Balance);
 
   const { appConfig } = useContext(AppConfigContext);
+
+  const [listBalance, setListBalance] = useState<Results<Balance>>(fullListBalance);
+
+  const setCurrentBalanceList = () => {
+    appConfig && setListBalance(fullListBalance.filtered('dueDate == $0', appConfig.dateToRenewBalance));
+  }
 
   const formatBalanceList = (): Array<BalanceHistoryItemInterface> => {
     const tempList: Array<BalanceHistoryItemInterface> = [];
@@ -50,6 +57,8 @@ const useBalance = () => {
   return {
     listCategory,
     listBalance,
+    setListBalance,
+    setCurrentBalanceList,
     formatBalanceList,
     getBalanceFromData,
     getAllCurrentValues