+++ app/models/interfaces/Filter.ts
@@ -1,0 +1,6 @@
+export default interface Filter {
+  transactionName?: string;
+  month?: string;
+  year?: string;
+  category?: string;
+}