+++ app/screens/filterTransaction/styles.ts
@@ -1,0 +1,13 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+  flatList: {
+    marginBottom: 80
+  },
+  infoMessage: {
+    fontSize: 20,
+    color: '#9a9b9c',
+    textAlign: 'center',
+    marginTop: 20,
+  }
+});